Add 0.2.21 release notes
